refactor: extract interface
