+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/CreatedRepository.java
@@ -2,7 +2,7 @@ package org.kinotic.github.internal.api.services.client;
 
 /**
  * Subset of GitHub's repository JSON returned by
- * {@link GitHubApiClient#createRepoFromTemplate}. Carries the fields the
+ * {@link DefaultGitHubApiClient#createRepoFromTemplate}. Carries the fields the
  * provisioner stamps onto the new {@code Project}.
  */
 public record CreatedRepository(Long id, String fullName, String defaultBranch) {}
