+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/GitHubApiException.java
@@ -1,7 +1,7 @@
 package org.kinotic.github.internal.api.services.client;
 
 /**
- * Thrown by {@link GitHubApiClient} for any non-success response from GitHub. Carries
+ * Thrown by {@link DefaultGitHubApiClient} for any non-success response from GitHub. Carries
  * the operation name plus status code and response body in the message.
  */
 public class GitHubApiException extends RuntimeException {
