+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/InstallationDetails.java
@@ -3,6 +3,6 @@ package org.kinotic.github.internal.api.services.client;
 /**
  * Subset of GitHub's installation JSON the platform reads — the install id plus
  * the owning account's login + type ({@code User} or {@code Organization}).
- * Returned by {@link GitHubApiClient#getInstallation}.
+ * Returned by {@link DefaultGitHubApiClient#getInstallation}.
  */
 public record InstallationDetails(Long id, String accountLogin, String accountType) {}
